Export Corregido con sus excepciones

+++ TP_FINAL_ALGORITMO/src/TP_Integrador_tmp/CsvExport.java
@@ -2,17 +2,25 @@ package TP_Integrador_tmp;
 
 import java.io.FileWriter;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 
 public class CsvExport extends CsvPrinter {
 
     public static void exportarComoCSV(DataFrame df, String rutaArchivo) {
 
-        if (df.isEmpty()) {
-            System.out.println("No hay datos para exportar.");
+        if (df.isEmpty()) { // Verifica si el df esta vacio
+            System.out.println("No hay datos dentro del DataFrame para exportar.");
            return;
         }
 
-        try (FileWriter writer = new FileWriter(rutaArchivo)) {
+        Path path = Paths.get(rutaArchivo); // Verifica si la rutaArchivo es valida
+        if (!Files.exists(path)) {
+            System.out.println("La ruta de archivo no es válida.");
+            return;
+        } else {
+            try (FileWriter writer = new FileWriter(rutaArchivo)) { // Usamos FileWriter
            
             // Escribir el encabezado
             int columnCount = df.getNroColumnas();
@@ -40,10 +48,18 @@ public class CsvExport extends CsvPrinter {
                     }
                 }
                 writer.append("\n");
+
             }
 
-        } catch (IOException e) {
-            e.printStackTrace();
+            System.out.println("Tu nuevo DataFrame se guardó correctamente en la ruta de archivo seleccionada");
+
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+
+
         }
+
+        
     }
 }
